api-gateway/requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
PyJWT==2.6.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
